feat: centralize label shortening
