+++ js/labelShortener.js
@@ -1,0 +1,34 @@
+// Utility for shortening verbose kink or subcategory labels
+
+// Map of specific long-form labels to their shortened versions
+const shortLabels = {
+  "Choosing my partner's outfit for the day or a scene": 'Choosing outfit',
+  "Selecting their underwear, lingerie or base layers": 'Picking underwear',
+  "Styling their hair (braiding, brushing, tying, etc.)": 'Styling hair',
+  "Picking head coverings (veils, hoods, hats)": 'Headwear',
+  "Offering makeup, polish, or accessories": 'Makeup/accessories',
+  "Creating themed looks (slutty, innocent, etc)": 'Themed looks',
+  "Dressing them in role-specific costumes": 'Roleplay outfits',
+  "Curating time-period or historical outfits": 'Historical outfits',
+  "Helping them present more femme/masc": 'Femme/masc styling',
+  "Coordinating their look with mine": 'Matching outfits',
+  "Implementing a 'dress ritual' or aesthetic preparation": 'Dress ritual',
+  "Enforcing a visual protocol (e.g. no bra, pigtails)": 'Visual protocol',
+  "Having my outfit selected for me by partner": 'Partner-picked outfit',
+  "Wearing chosen lingerie/underwear": 'Chosen lingerie',
+  "Having my hair brushed, braided, styled": 'Hair styled by partner'
+};
+
+/**
+ * Shorten a label for use in compact layouts.
+ * Falls back to the first four words of the label when no mapping exists.
+ * @param {string} text
+ * @returns {string}
+ */
+export function shortenLabel(text = '') {
+  if (!text) return '';
+  return shortLabels[text] || text.split(/\s+/).slice(0, 4).join(' ');
+}
+
+export default shortenLabel;
+
